libqmi-glib,enums-dms: fix warnings with -Wsign-compare
  qmi-enums-dms.c: In function ‘qmi_dms_swi_usb_composition_get_description’:
  qmi-enums-dms.c:58:67: warning: comparison of integer expressions of different signedness: ‘QmiDmsSwiUsbComposition’ {aka ‘enum <anonymous>’} and ‘long unsigned int’ [-Wsign-compare]
     58 |     return ((value > QMI_DMS_SWI_USB_COMPOSITION_UNKNOWN && value < G_N_ELEMENTS (usb_composition_description)) ? usb_composition_description[value] : NULL);
        |                                                                   ^

+++ src/libqmi-glib/qmi-enums-dms.c
@@ -55,5 +55,7 @@ static const gchar *usb_composition_description[] = {
 const gchar *
 qmi_dms_swi_usb_composition_get_description (QmiDmsSwiUsbComposition value)
 {
-    return ((value > QMI_DMS_SWI_USB_COMPOSITION_UNKNOWN && value < G_N_ELEMENTS (usb_composition_description)) ? usb_composition_description[value] : NULL);
+    return (((value > QMI_DMS_SWI_USB_COMPOSITION_UNKNOWN) &&
+             ((guint)value < G_N_ELEMENTS (usb_composition_description))) ?
+            usb_composition_description[value] : NULL);
 }
